Move DownloadDataDialog to RightSidebar folder
